+++ packages/dialect-react/CHANGELOG.md
@@ -1,6 +1,7 @@
 # CHANGELOG
 
 ## [UNRELEASED]
+- fix: added hash value for array, arrays comparing by hash not by message length
 
 ## [0.2.0] - 2022-03-31